Fix CA1819 in generated code
Arrays returned by properties are not write-protected, even if the property is read-only. To keep the array tamper-proof, the property must return a copy of the array. To avoid copying on array access generated code now uses the ImmutableArray type.

ImmutableArray usually requires copying the source array, however we use Unsafe methods to skip that too.