Fix overflow issue leading to failed opcode lookup
Casting unsigned opcodes to a signed integer was causing overflow issues that would turn values over 2,147,483,647 into a bit pattern interpreted as a signed integer. after that any subsequent lookups of the original unsigned opcode value would fail.

To fix this we now use a nullable type for BebopRecord rather than using '-1' to signify that a record didn't have one.